category tabs - sx update
